:recycle: Move helpers into routes mod
