+++ src/client/mock_server/routes/mod.rs
@@ -1,9 +1,9 @@
 pub mod auth;
 pub mod token;
 
-use errors::Result;
+use client::params::UrlQueryParams;
+use errors::{Error, Result};
 use url::Url;
-use client::mock_server::mock_server::single_param;
 
 pub use self::auth::auth_response as auth_route;
 pub use self::token::token_response as token_route;
@@ -13,6 +13,30 @@ mod auth_spec;
 #[cfg(test)]
 mod token_spec;
 
+#[inline]
 pub fn parse_state(url: &Url) -> Result<String> {
     single_param("state", url)
+}
+
+#[inline]
+pub fn maybe_single_param(name: &'static str, url: &Url) -> Option<String> {
+    match UrlQueryParams::from(url.query_pairs()).get(name) {
+        Some(v) => v.single().map(|v| v.clone()),
+        None => None,
+    }
+}
+
+#[inline]
+pub fn single_param(name: &'static str, url: &Url) -> Result<String> {
+    match UrlQueryParams::from(url.query_pairs()).get(name) {
+        Some(v) => Ok(v.single()
+            .ok_or(Error::msg(
+                "Bad Request: Expected Single Parameter, found many",
+            ))?
+            .clone()),
+        None => Err(Error::invalid_request(
+            Some(format!("Bad Request: Missing `{}`", name)),
+            None,
+        )),
+    }
 }
